Rename fox-info.js to fox.js and remove all the Redis connection warning message
